refactor: ModalWrapper 컴포넌트 분리
