+++ src/provider/ModalProvider/ModalWrapper.tsx
@@ -1,0 +1,90 @@
+import { useEffect, useId, useRef } from 'react';
+import { cn } from '@/lib/utils';
+import { ModalContext } from './ModalContext';
+
+/** 모달 스택에 저장되는 모달 정보 */
+export type Modal = {
+  id: string;
+  render: () => React.ReactNode;
+};
+
+/** 개별 모달 내부에서 사용하는 Context 값 */
+export interface ModalContextValue {
+  /** aria-labelledby 연결용 ID */
+  labelId: string;
+  /** aria-describedby 연결용 ID */
+  descriptionId: string;
+  /** 현재 모달 닫기 */
+  close: () => void;
+}
+
+interface ModalWrapperProps {
+  modal: Modal;
+  isTop: boolean;
+  close: () => void;
+}
+
+/**
+ * 개별 모달을 <dialog>로 감싸는 Wrapper
+ * - 브라우저 네이티브 Focus trap, Top Layer 사용
+ * - 접근성 속성 (aria-modal, aria-labelledby, aria-describedby) 제공
+ * - ESC 키, backdrop 클릭 처리
+ */
+export default function ModalWrapper({
+  modal,
+  isTop,
+  close,
+}: ModalWrapperProps) {
+  const dialogRef = useRef<HTMLDialogElement>(null);
+  const labelId = useId();
+  const descriptionId = useId();
+
+  useEffect(() => {
+    const dialog = dialogRef.current;
+    if (dialog && !dialog.open) {
+      dialog.showModal();
+    }
+  }, []);
+
+  const handleCancel = (e: React.SyntheticEvent<HTMLDialogElement>) => {
+    e.preventDefault();
+    if (isTop) {
+      close();
+    }
+  };
+
+  const handleClick = (e: React.MouseEvent<HTMLDialogElement>) => {
+    if (e.target === e.currentTarget && isTop) {
+      close();
+    }
+  };
+
+  const contextValue: ModalContextValue = {
+    labelId,
+    descriptionId,
+    close,
+  };
+
+  return (
+    <dialog
+      ref={dialogRef}
+      className={cn(
+        'fixed inset-0 m-0 h-dvh w-dvw bg-transparent p-0',
+        'flex items-center justify-center outline-none',
+        isTop
+          ? 'backdrop:animate-in backdrop:fade-in-0 backdrop:bg-black/50'
+          : 'backdrop:bg-transparent',
+        !isTop && 'pointer-events-none *:pointer-events-auto'
+      )}
+      aria-modal="true"
+      aria-labelledby={labelId}
+      aria-describedby={descriptionId}
+      onCancel={handleCancel}
+      onClick={handleClick}
+    >
+      <ModalContext.Provider value={contextValue}>
+        {modal.render()}
+      </ModalContext.Provider>
+    </dialog>
+  );
+}
